+++ apps/admin-x-settings/src/components/settings/growth/GrowthSettings.tsx
@@ -3,13 +3,13 @@ import Offers from './Offers';
 import React from 'react';
 import Recommendations from './Recommendations';
 import SearchableSection from '../../SearchableSection';
-import TipsOrDonations from './TipsOrDonations';
+import TipsAndDonations from './TipsAndDonations';
 import useFeatureFlag from '../../../hooks/useFeatureFlag';
 import {checkStripeEnabled} from '@tryghost/admin-x-framework/api/settings';
 import {useGlobalData} from '../../providers/GlobalDataProvider';
 
 export const searchKeywords = {
-    tips: ['growth', 'tip', 'donation', 'one time', 'payment'],
+    tips: ['growth', 'tips', 'donations', 'one time', 'payment'],
     embedSignupForm: ['growth', 'embeddable signup form', 'embeddable form', 'embeddable sign up form', 'embeddable sign up'],
     recommendations: ['growth', 'recommendations', 'recommend', 'blogroll'],
     offers: ['growth', 'offers', 'discounts', 'coupons', 'promotions']
@@ -25,7 +25,7 @@ const GrowthSettings: React.FC = () => {
             <Recommendations keywords={searchKeywords.recommendations} />
             <EmbedSignupForm keywords={searchKeywords.embedSignupForm} />
             {hasStripeEnabled && <Offers keywords={searchKeywords.offers} />}
-            {hasTipsAndDonations && <TipsOrDonations keywords={searchKeywords.tips} />}
+            {hasTipsAndDonations && <TipsAndDonations keywords={searchKeywords.tips} />}
         </SearchableSection>
     );
 };
